Removed ShellMiddleware, added shell_stream in views.py
shell.py contains the Shell class which is used for any kind
of SSH commands. It can run commands, optionally parsing the
output, has the ability to automatically configure itself for a
certain machine (find right key, user etc).

shell_stream parses the output using the hidden iframe technique
as ShellMiddleware used to do.

keep in mind that the docstrings in several functions and
class methods are outdated.