Added the tests config file (yaml)
We need to replicate a user, so we should not seek in our
actual database for information and rather compare the
information that the api gives with the actual.

We use a yaml file to store information received from our
API, much like somebody using the API would do...

IMPORTANT:
You need to add some things to yaml file prior to running
nosetests. See our wiki page for more information.