Make graphs interrupt line in null ranges
We send null values when we have no monitoring data available
and this will make the graphs interrupt the line in the no-data
ranges instead of showing false 0 values (The line will simply
be undefined in those ranges). Apart from making obviously
distinguishable when there are 0 values from when there are
actually no available data, this will save us from anomalies
in the range boundaries concerning metrics calculated with use
of derivatives.