Closing shell: disconnect namespace, not socket
Mist and Shell socketio namespaces are multiplexed over a single
websocket connection. When closing the shell, we want to disconnect
from the shell namespace, not disconnect the websocket. Currently,
closing the shell disconnects the websocket which is then immediately
reconnecting (with a new session id). This commit fixes this behavor,
and subsequent shells will connect using the same websocket connection.