Create DNS A record on machine provisioning
If 'hostname' param is passed on machine creation POST, then all
backends for which there's DNS support will be searched for the
appropriate DNS zone, and if one is found, then an A record will be
created for the supplied hostname and the newly create machine's public
IP address.

So far only EC2 backends are supported.