Use non blocking requests in sockjs-tornado
Use the trequests library, which in turn uses tornalet (something like
tornado greenlets), to patch requests into usign a non blocking HTTP
adapter.

This is done in order to avoid blocking tornado ioloop with HTTP calls
performed in the context of sockjs-tornado, using the requests library.
The main, if not only, place we do this, is when performing get_stats
requests to the monitoring component, in order to send the results to
the client over the websocket.