Split controllers into smaller entities
This MR introduces a minor refactor in our controllers. The
`BaseController` no longer holds the entire functionality

The `BaseController` includes very basic cloud actions, such as add,
delete, and edit. It also initiates various subcontrollers as
attributes

The subcontrollers are responsible for holding the entire functionality
of a Mist.io Cloud as it is offered by libcloud